pyproject.toml
[project]
name = "app-customers"
version = "0.1.0"
description = ""
authors = [
    {name = "derekSaga",email = "derek.coelho@outlook.com.br"}
]
readme = "README.md"
requires-python = ">=3.13.5, <=3.14.2"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "ruff (>=0.14.13,<0.15.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pytestarch (>=4.0.1,<5.0.0)",
    "sqlalchemy[asyncio] (>=2.0.45,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "redis (>=7.1.0,<8.0.0)",
    "google-cloud-pubsub (>=2.34.0,<3.0.0)",
    "asgi-correlation-id (>=4.3.4,<5.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [

    "pytest (>=8.0.0,<9.0.0)",

    "pytest-coverage (>=0.0,<0.1)"
]

[tool.ruff]
line-length = 79
target-version = "py313"

[tool.ruff.lint]
# E: Pycodestyle, F: Pyflakes, I: Isort (imports), UP: Pyupgrade (modernizar sintaxe)
select = ["E", "F", "I", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.coverage.report]
fail_under = 85
